+++ src/Tuple.php
@@ -25,13 +25,7 @@ use function method_exists;
 final class Tuple implements Countable, IteratorAggregate
 {
     /**
-     * The tuple collection.
-     *
-     * Each tuple is an array with two value.
-     * The first value at index 0 represents the label name and is a string.
-     * The second value at index 1 represents a Period or a Sequence object.
-     *
-     * @var array<int, array<int, Period|Sequence|string>>.
+     * @var array<int, array{0:string, 1:Sequence|Period}>.
      */
     private $pairs = [];
 
@@ -41,7 +35,7 @@ final class Tuple implements Countable, IteratorAggregate
     public function __construct(iterable $pairs = [])
     {
         foreach ($pairs as [$offset, $value]) {
-            $this->addPair($offset, $value);
+            $this->append($offset, $value);
         }
     }
 
@@ -57,35 +51,23 @@ final class Tuple implements Countable, IteratorAggregate
 
         $labels = $labelGenerator->generate($sequence);
         foreach ($sequence as $offset => $period) {
-            $tuple->addPair($labels[$offset], $period);
+            $tuple->append($labels[$offset], $period);
         }
 
         return $tuple;
     }
 
     /**
-     * A a new pair to the collection.
-     *
-     * @param mixed $label any stringable structure.
-     * @param mixed $input if the input is not a Period or a Sequence instance it is not added.
+     * Creates a new collection from a generic iterable structure.
      */
-    public function addPair($label, $input): void
+    public static function fromIterable(iterable $iterable): self
     {
-        if (null !== $label && !is_scalar($label) && !method_exists($label, '__toString')) {
-            return;
+        $tuple = new self();
+        foreach ($iterable as $offset => $value) {
+            $tuple->append($offset, $value);
         }
 
-        if ($input instanceof Period || $input instanceof Sequence) {
-            $this->pairs[] = [(string) $label, $input];
-        }
-    }
-
-    /**
-     * Tells whether the collection is empty.
-     */
-    public function isEmpty(): bool
-    {
-        return [] === $this->pairs;
+        return $tuple;
     }
 
     /**
@@ -107,6 +89,14 @@ final class Tuple implements Countable, IteratorAggregate
     }
 
     /**
+     * Tells whether the collection is empty.
+     */
+    public function isEmpty(): bool
+    {
+        return [] === $this->pairs;
+    }
+
+    /**
      * Returns the collection boundaries.
      */
     public function boundaries(): ?Period
@@ -115,11 +105,28 @@ final class Tuple implements Countable, IteratorAggregate
         foreach ($this->pairs as [$name, $item]) {
             if ($item instanceof Period) {
                 $sequence->push($item);
-            } elseif ($item instanceof Sequence) {
-                $sequence->push(...$item);
+                continue;
             }
+            $sequence->push(...$item);
         }
 
         return $sequence->boundaries();
+    }
+
+    /**
+     * Add a new pair to the collection.
+     *
+     * @param mixed $label any stringable structure.
+     * @param mixed $input if the input is not a Period or a Sequence instance it is not added.
+     */
+    public function append($label, $input): void
+    {
+        if (!is_scalar($label) && !method_exists($label, '__toString')) {
+            return;
+        }
+
+        if ($input instanceof Period || $input instanceof Sequence) {
+            $this->pairs[] = [(string) $label, $input];
+        }
     }
 }