+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All Notable changes to `Period Visualizer` will be documented in this file
 
 # 0.5.0 - TBD
 
+- Added the `Tuple` class.
 - Added the `Writer` interface and the `Stdout` implementation, output is done via `fwrite` instead of `echo`
 - Added the `Viewer::view` method as the most generic method
 - Added Decimal zero fill left padding to `DecimalNumber` LabelGenerator implementing class